Add send_cmd (w) support
You can do something like "w slice set 0 rxant=RX_B" and the command
will get sent through to the Flex and you'll get the hex result code as
a response.

It's up to you to know what you're doing, it's up to you to figure out
things like slice number mappings on your own, and if you change state
behind nCAT's back it might give incorrect responses to hamlib commands
in the future (the flex protocol broadcasts most state changes to
everyone *except* the client that requested the change, on the
assumption that that client knows what it's doing, raw_cmd is a
deliberate layering violation so we *don't* know what we're doing).
Basically only use it for things that are outside of the hamlib API.