Add a basic \get_powerstat command
Recent hamlib seems to treat an error on get_powerstat as equivalent to
a 0 return ("power is off"), and then complains that it can't do
operations on a rig that's off.

Add an implementation of get_powerstat that always returns 1 (power on),
since nCAT can't be connected to a rig that's off.

Theoretically we could support set_powerstat 0 to send a shutdown
command, but nah.