Add "protocol 1" dumpstate support
Add flags saying that we don't support the set_vfo command (the stub
implementation is still there for clients that don't understand the flag
or don't care). There's some code in rig_open that tries to read the
VFOB frequency (without checking the list of VFOs first), but telling it
that we can't set_vfo prevents it from sending the command and
generating a warning in the nCAT logs.

We could also send rig_model=RIG_MODEL_F6K here, but that could be more
trouble than it's worth if something in the future decides that it means
we have some quirk of the hamlib-FlexCAT-kenwood emulation that we
actually don't.