Add experimental metering support
When enabled, hamlib clients to request S-meter, Po, SWR, ALC,
compression, PA voltage, and PA temperature using the "l" command
(clients that can make use of this include JTDX and grig).

There may be bugs, and enabling results in about 20kbit/s of additional
UDP traffic from the radio as long as nCAT is connected, so it's off by
default.

In the future, I plan on adding a way to pay the traffic cost only if a
client actually uses metering, and then the default will change to on,
but it will still be possible to disable it (which will turn off
advertising those levels in the rig caps).