big command-handling refactor
Make handlers aware of their names (formerly names were only passed to
AddHandler).

Promote the "long name" to a first class citizen -- every handler has
one long name and zero or one short names, instead of an array of names.

The backslash is no longer part of the long name, it's parsed out separately.

All of this will, with a little more annotating, allow supporting the
hamlib "extended protocol", in which the first thing a command has to
send back in a response is its own long name (whether invoked via the
long or short name).